+++ navigator/r2-navigator-swift/EPUB/Assets/Static/scripts/readium-reflowable.js
@@ -4704,6 +4704,7 @@ function log() {
 "use strict";
 __webpack_require__.r(__webpack_exports__);
 /* harmony export */ __webpack_require__.d(__webpack_exports__, {
+/* harmony export */   "getColumnCountPerScreen": () => (/* binding */ getColumnCountPerScreen),
 /* harmony export */   "isScrollModeEnabled": () => (/* binding */ isScrollModeEnabled),
 /* harmony export */   "scrollToId": () => (/* binding */ scrollToId),
 /* harmony export */   "scrollToPosition": () => (/* binding */ scrollToPosition),
@@ -4746,6 +4747,11 @@ window.addEventListener(
 window.addEventListener(
   "load",
   function () {
+    const observer = new ResizeObserver(() => {
+      appendVirtualColumnIfNeeded();
+    });
+    observer.observe(document.body);
+
     // on page load
     window.addEventListener("orientationchange", function () {
       orientationChanged();
@@ -4755,6 +4761,34 @@ window.addEventListener(
   },
   false
 );
+
+/**
+ * Having an odd number of columns when displaying two columns per screen causes snapping and page
+ * turning issues. To fix this, we insert a blank virtual column at the end of the resource.
+ */
+function appendVirtualColumnIfNeeded() {
+  const id = "readium-virtual-page";
+  var virtualCol = document.getElementById(id);
+  if (isScrollModeEnabled() || getColumnCountPerScreen() != 2) {
+    virtualCol?.remove();
+  } else {
+    var documentWidth = document.scrollingElement.scrollWidth;
+    var pageWidth = window.innerWidth;
+    var colCount = documentWidth / pageWidth;
+    var hasOddColCount = (Math.round(colCount * 2) / 2) % 1 > 0.1;
+    if (hasOddColCount) {
+      if (virtualCol) {
+        virtualCol.remove();
+      } else {
+        virtualCol = document.createElement("div");
+        virtualCol.setAttribute("id", id);
+        virtualCol.style.breakBefore = "column";
+        virtualCol.innerHTML = "&#8203;"; // zero-width space
+        document.body.appendChild(virtualCol);
+      }
+    }
+  }
+}
 
 var last_known_scrollX_position = 0;
 var last_known_scrollY_position = 0;
@@ -4808,6 +4842,14 @@ function orientationChanged() {
     window.orientation === 0 || window.orientation == 180
       ? screen.width
       : screen.height;
+}
+
+function getColumnCountPerScreen() {
+  return parseInt(
+    window
+      .getComputedStyle(document.documentElement)
+      .getPropertyValue("column-count")
+  );
 }
 
 function isScrollModeEnabled() {