+++ navigator/CHANGELOG.md
@@ -11,6 +11,10 @@ All notable changes to this project will be documented in this file.
 
 * A new `translate` EPUB and PDF editing action is available for iOS 15.
 
+### Fixed
+
+* Fixed turning pages of an EPUB reflowable resource with an odd number of columns. A virtual blank trailing column is appended to the resource when displayed as two columns.
+
 
 ## [2.1.0]